Started by cleaning up the table tests. This lead to a speed improvement: these tests were 40 seconds, now 16. Speed improvement gained by using new method "uses_page" instead of "use_page": we'll need to use this in other tests as well!
However, getting this to work required a small fix to $htmlRoot which complicated other test code which i eventually determined was need to workaround a bug in IE#url_list, which was being stored in the wrong format for "file" hrefs (a real corner case, but significant for our unit tests). Fixing these ended up allowing me to remove a fair amount of noisy code from other tests. 

git-svn-id: https://svn.openqa.org/svn/watir/trunk@1342 deaf61b0-630d-0410-ad59-fc55e6353476