This is Jira 233 contributed by Aidy Lewis and Michael Hwee. This adds multiple attribute support to FireWatir.
I have verified that the mozilla_all_tests.rb runs and that therefore this does not introduce regression bugs that this test suite protects us from.

I have not verified the new functionality, but will do that soon using Watir's core tests.

Note that Michael made a valiant effort, in his original submission, to place all the FireWatir classes in the FireWatir module. Right idea, but i was not able to get the tests to pass and therefore backed this out.

Thanks Aidy and Michael.

git-svn-id: https://svn.openqa.org/svn/watir/trunk@1439 deaf61b0-630d-0410-ad59-fc55e6353476